+++ ChangeLog.md
@@ -8,13 +8,15 @@ to [Semantic Versioning](http://semver.org/) rules.
 
 ### Added
 
+- Added SnowthClient.RebuildActivity() to request a rebuild of IRONdb activity
+tracking data for a list of metrics supplied in the new type RebuildActivityRequest.
 - Added SnowthClient.WriteRawMetricList() convenience function to support writing
 raw metric data directly to IRONdb via FlatBuffers Objects.
 
 ### Changed
 
 - Changed the signature for SnowthClient.WriteRaw() to return the status
-result of the /raw operation with the new type WriteRawResponse.
+result of the /raw operation with the new type IRONdbPutResponse.
 
 ### Fixed